+++ attribute/hash.go
@@ -27,6 +27,7 @@ const (
 	int64SliceID   uint64 = 3762322556277578591 // "_[]int64" (little endian)
 	float64SliceID uint64 = 7308324551835016539 // "[]double" (little endian)
 	stringSliceID  uint64 = 7453010373645655387 // "[]string" (little endian)
+	sliceID        uint64 = 7883494272577650031 // "__slice_" (little endian)
 )
 
 // hashKVs returns a new xxHash64 hash of kvs.
@@ -79,6 +80,13 @@ func hashKV(h xxhash.Hash, kv KeyValue) xxhash.Hash {
 		rv := reflect.ValueOf(kv.Value.slice)
 		for i := 0; i < rv.Len(); i++ {
 			h = h.String(rv.Index(i).String())
+		}
+	case SLICE:
+		h = h.Uint64(sliceID)
+		rv := reflect.ValueOf(kv.Value.slice)
+		for i := 0; i < rv.Len(); i++ {
+			v := rv.Index(i).Interface().(Value)
+			h = hashKV(h, KeyValue{Key: "", Value: v})
 		}
 	case INVALID:
 	default: