+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The next release will require at least [Go 1.25].
 ### Added
 
 - Support testing of [Go 1.26]. (#7902)
+- Add `Slice` and `SliceValue` functions for new `SLICE` Type in `go.opentelemetry.io/otel/attribute`.
+  This adds support for heterogeneous arrays as attribute values. (#7944)
 
 ### Fixed